Began an hp/etc info bar
-Added the SDL2_TTF library to the project
-Added basic methods/fields to entity/player for tracking of health
-Successfully added a crude display of player's current health

+++ Game Engine/Player.h
@@ -4,7 +4,7 @@ class Player :
 	public Character
 {
 public:
-	Player(int h, int w, double positionX, double positionY, double velX, double velY, int hp);
+	Player(int h, int w, double positionX, double positionY, double velX, double velY, double hp);
 	~Player();
 	void determineMovement(double playerPosX, double plaerPosY) override;
 	void setHorizontalMovementKeyPress(short direction);
